refactor: moving Splice into splice.rs
